requirements.txt
Django==5.2.5
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
djangorestframework-xml==2.0.0
djoser==2.3.3
drf-yasg==1.21.10
django-prometheus==2.4.1
daphne[http2,tls]==4.2.1
gradio==5.42.0
debugpy==1.8.16
matplotlib==3.10.5
# flash_attn==2.5.8
numpy==1.24.4
Pillow==11.3.0
Requests==2.32.4
torch==2.8.0
torchvision==0.23.0
transformers==4.55.2
bitsandbytes==0.47.0
python-dotenv==1.1.1
gunicorn==23.0.0
scikit-build-core==0.11.5
cmake==4.1.0
ninja==1.13.0
dill==0.4.0
psycopg2-binary==2.9.10
# git+https://github.com/huggingface/transformers
pdfplumber==0.11.7
pytesseract==0.3.13
unstructured.pytesseract==0.3.15
pdf2image==1.17.0
django-countries==7.6.1
# faiss-gpu==1.7.2
faiss-cpu
pi-heif==1.1.0
tensorboard==2.20.0
colorama==0.4.6
polars==1.32.3
selenium==4.35.0
webdriver-manager==4.0.2
googlesearch-python==1.3.0
twisted[http2,tls]==25.5.0
ray[data,train,tune,serve,default,rllib]==2.48.0
validators==0.35.0
python-decouple==3.8
uvicorn
fastapi
pydantic-settings
sse-starlette
starlette-context
